refactor: useAuth 관련 react-query 재정의
