--- src/query-hooks/useAuth/POST/index.ts
@@ -1,1 +1,0 @@
-export { default } from './usePostAuth';
