+++ src/query-hooks/useAuth/index.ts
@@ -1,5 +1,31 @@
-import POST from './POST';
+import { AxiosError } from 'axios';
+import { useMutation } from 'react-query';
 
-export default {
-    POST
+// apis
+import { postAuthenticateUser, postLogoutUser } from './api';
+
+import type { Post } from './api.type';
+
+// [POST] 로그인 인증 요청
+const usePostAuthenticateUser = () => {
+    return useMutation<
+        Post.LoginAuthenticate.Return,
+        AxiosError,
+        Post.LoginAuthenticate.ReqParams
+    >(({ email, password }) => postAuthenticateUser({ email, password }), {
+        onSuccess: (data: Post.LoginAuthenticate.Return) => {
+            return Promise.resolve(data);
+        }
+    });
 };
+
+// [POST] 로그아웃 요청
+const usePostLogoutUser = () => {
+    return useMutation<Post.Logout.Return, AxiosError>(() => postLogoutUser(), {
+        onSuccess: (data: Post.Logout.Return) => {
+            return Promise.resolve(data);
+        }
+    });
+};
+
+export { usePostAuthenticateUser, usePostLogoutUser };
